chore: Bookmark 패키지 분리
