+++ backend/src/docs/asciidoc/index.adoc
@@ -11,3 +11,4 @@ toc::[]
 include::topic.adoc[]
 include::pin.adoc[]
 include::member.adoc[]
+include::bookmark.adoc[]
