+++ backend/src/main/java/com/mapbefine/mapbefine/bookmark/presentation/BookmarkController.java
@@ -1,0 +1,76 @@
+package com.mapbefine.mapbefine.bookmark.presentation;
+
+import com.mapbefine.mapbefine.auth.domain.AuthMember;
+import com.mapbefine.mapbefine.bookmark.application.BookmarkCommandService;
+import com.mapbefine.mapbefine.bookmark.application.BookmarkQueryService;
+import com.mapbefine.mapbefine.bookmark.dto.response.BookmarkResponse;
+import com.mapbefine.mapbefine.common.interceptor.LoginRequired;
+import java.net.URI;
+import java.util.List;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/members/bookmarks")
+public class BookmarkController {
+
+    private final BookmarkCommandService bookmarkCommandService;
+
+    private final BookmarkQueryService bookmarkQueryService;
+
+    public BookmarkController(
+            BookmarkCommandService bookmarkCommandService,
+            BookmarkQueryService bookmarkQueryService
+    ) {
+        this.bookmarkCommandService = bookmarkCommandService;
+        this.bookmarkQueryService = bookmarkQueryService;
+    }
+
+    @LoginRequired
+    @PostMapping
+    public ResponseEntity<Void> addTopicInBookmark(
+            AuthMember authMember,
+            @RequestParam Long topicId
+    ) {
+        Long bookmarkId = bookmarkCommandService.addTopicInBookmark(authMember, topicId);
+
+        return ResponseEntity.created(URI.create("/members/bookmarks/" + bookmarkId)).build();
+    }
+
+    @LoginRequired
+    @GetMapping
+    public ResponseEntity<List<BookmarkResponse>> findAllTopicsInBookmark(
+            AuthMember authMember
+    ) {
+        List<BookmarkResponse> topicsInBookmark
+                = bookmarkQueryService.findAllTopicsInBookmark(authMember);
+
+        return ResponseEntity.ok(topicsInBookmark);
+    }
+
+    @LoginRequired
+    @DeleteMapping("{bookmarkId}")
+    public ResponseEntity<Void> deleteTopicInBookmark(
+            AuthMember authMember,
+            @PathVariable Long bookmarkId
+    ) {
+        bookmarkCommandService.deleteTopicInBookmark(authMember, bookmarkId);
+
+        return ResponseEntity.noContent().build();
+    }
+
+    @LoginRequired
+    @DeleteMapping
+    public ResponseEntity<Void> deleteAllTopicsInBookmark(AuthMember authMember) {
+        bookmarkCommandService.deleteAllBookmarks(authMember);
+
+        return ResponseEntity.noContent().build();
+    }
+
+}
