+++ backend/src/main/java/com/mapbefine/mapbefine/member/application/MemberCommandService.java
@@ -3,8 +3,6 @@ package com.mapbefine.mapbefine.member.application;
 import com.mapbefine.mapbefine.auth.domain.AuthMember;
 import com.mapbefine.mapbefine.member.domain.Member;
 import com.mapbefine.mapbefine.member.domain.MemberRepository;
-import com.mapbefine.mapbefine.member.domain.MemberTopicBookmark;
-import com.mapbefine.mapbefine.member.domain.MemberTopicBookmarkRepository;
 import com.mapbefine.mapbefine.member.domain.MemberTopicPermission;
 import com.mapbefine.mapbefine.member.domain.MemberTopicPermissionRepository;
 import com.mapbefine.mapbefine.member.dto.request.MemberCreateRequest;
@@ -23,18 +21,14 @@ public class MemberCommandService {
     private final TopicRepository topicRepository;
     private final MemberTopicPermissionRepository memberTopicPermissionRepository;
 
-    private final MemberTopicBookmarkRepository memberTopicBookmarkRepository;
-
     public MemberCommandService(
             MemberRepository memberRepository,
             TopicRepository topicRepository,
-            MemberTopicPermissionRepository memberTopicPermissionRepository,
-            MemberTopicBookmarkRepository memberTopicBookmarkRepository
+            MemberTopicPermissionRepository memberTopicPermissionRepository
     ) {
         this.memberRepository = memberRepository;
         this.topicRepository = topicRepository;
         this.memberTopicPermissionRepository = memberTopicPermissionRepository;
-        this.memberTopicBookmarkRepository = memberTopicBookmarkRepository;
     }
 
     public Long save(MemberCreateRequest request) {
@@ -121,60 +115,6 @@ public class MemberCommandService {
         validateMemberCanTopicUpdate(authMember, memberTopicPermission.getTopic());
 
         memberTopicPermissionRepository.delete(memberTopicPermission);
-    }
-
-    public Long addTopicInBookmark(AuthMember authMember, Long topicId) {
-        Topic topic = getTopicById(topicId);
-        validateBookmarkingPermission(authMember, topic);
-        Member member = getMemberById(authMember);
-
-        MemberTopicBookmark memberTopicBookmark
-                = MemberTopicBookmark.createWithAssociatedTopicAndMember(topic, member);
-        memberTopicBookmarkRepository.save(memberTopicBookmark);
-
-        return memberTopicBookmark.getId();
-    }
-
-    private Member getMemberById(AuthMember authMember) {
-        return memberRepository.findById(authMember.getMemberId())
-                .orElseThrow(() -> new NoSuchElementException("존재하지 않는 멤버입니다."));
-    }
-
-    private Topic getTopicById(Long topicId) {
-        return topicRepository.findById(topicId)
-                .orElseThrow(() -> new NoSuchElementException("존재하지 않는 토픽입니다."));
-    }
-
-    private void validateBookmarkingPermission(AuthMember authMember, Topic topic) {
-        if (authMember.canRead(topic)) {
-            return;
-        }
-
-        throw new IllegalArgumentException("토픽에 대한 권한이 없어서 즐겨찾기에 추가할 수 없습니다.");
-    }
-
-    public void deleteTopicInBookmark(AuthMember authMember, Long bookmarkId) {
-        validateBookmarkDeletingPermission(authMember, bookmarkId);
-
-        memberTopicBookmarkRepository.deleteById(bookmarkId);
-
-    }
-
-    private void validateBookmarkDeletingPermission(AuthMember authMember, Long bookmarkId) {
-        boolean canDelete = memberTopicBookmarkRepository.existsByIdAndMemberId(
-                bookmarkId,
-                authMember.getMemberId(
-                ));
-
-        if (canDelete) {
-            return;
-        }
-
-        throw new IllegalArgumentException("즐겨찾기 삭제에 대한 권한이 없습니다.");
-    }
-
-    public void deleteAllBookmarks(AuthMember authMember) {
-        memberTopicBookmarkRepository.deleteAllByMemberId(authMember.getMemberId());
     }
 
 }
