+++ backend/src/test/java/com/mapbefine/mapbefine/member/application/MemberQueryServiceTest.java
@@ -6,6 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import com.mapbefine.mapbefine.auth.domain.AuthMember;
 import com.mapbefine.mapbefine.auth.domain.member.Guest;
 import com.mapbefine.mapbefine.auth.domain.member.User;
+import com.mapbefine.mapbefine.bookmark.domain.BookmarkRepository;
 import com.mapbefine.mapbefine.common.annotation.ServiceTest;
 import com.mapbefine.mapbefine.location.LocationFixture;
 import com.mapbefine.mapbefine.location.domain.Location;
@@ -13,14 +14,11 @@ import com.mapbefine.mapbefine.location.domain.LocationRepository;
 import com.mapbefine.mapbefine.member.MemberFixture;
 import com.mapbefine.mapbefine.member.domain.Member;
 import com.mapbefine.mapbefine.member.domain.MemberRepository;
-import com.mapbefine.mapbefine.member.domain.MemberTopicBookmark;
-import com.mapbefine.mapbefine.member.domain.MemberTopicBookmarkRepository;
 import com.mapbefine.mapbefine.member.domain.MemberTopicPermission;
 import com.mapbefine.mapbefine.member.domain.MemberTopicPermissionRepository;
 import com.mapbefine.mapbefine.member.domain.Role;
 import com.mapbefine.mapbefine.member.dto.response.MemberDetailResponse;
 import com.mapbefine.mapbefine.member.dto.response.MemberResponse;
-import com.mapbefine.mapbefine.member.dto.response.MemberTopicBookmarkResponse;
 import com.mapbefine.mapbefine.member.dto.response.MemberTopicPermissionDetailResponse;
 import com.mapbefine.mapbefine.member.dto.response.MemberTopicPermissionResponse;
 import com.mapbefine.mapbefine.pin.PinFixture;
@@ -59,7 +57,7 @@ class MemberQueryServiceTest {
     private PinRepository pinRepository;
 
     @Autowired
-    private MemberTopicBookmarkRepository memberTopicBookmarkRepository;
+    private BookmarkRepository bookmarkRepository;
 
     @Test
     @DisplayName("Topic 에 권한이 있는자들을 모두 조회한다.")
@@ -257,40 +255,6 @@ class MemberQueryServiceTest {
         // given when then
         assertThatThrownBy(() -> memberQueryService.findTopicsByMember(new Guest()))
                 .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    @DisplayName("즐겨찾기 목록에 추가 된 토픽을 조회할 수 있다")
-    public void findAllTopicsInBookmark_success() {
-        //given
-        Member creator = MemberFixture.create("creator", "member@naver.com", Role.USER);
-        memberRepository.save(creator);
-
-        Topic topic1 = topicRepository.save(TopicFixture.createPublicAndAllMembersTopic(creator));
-        Topic topic2 = topicRepository.save(TopicFixture.createPublicAndAllMembersTopic(creator));
-        topicRepository.save(topic1);
-        topicRepository.save(topic2);
-
-        //when
-        Member otherMember =
-                MemberFixture.create("otherMember", "otherMember@naver.com", Role.USER);
-        memberRepository.save(otherMember);
-        MemberTopicBookmark bookmarkingTopic1 =
-                MemberTopicBookmark.createWithAssociatedTopicAndMember(topic1, otherMember);
-        MemberTopicBookmark bookmarkingTopic2 =
-                MemberTopicBookmark.createWithAssociatedTopicAndMember(topic2, otherMember);
-
-        memberTopicBookmarkRepository.save(bookmarkingTopic1);
-        memberTopicBookmarkRepository.save(bookmarkingTopic2);
-
-        //then
-        List<MemberTopicBookmarkResponse> topicsInBookmark
-                = memberQueryService.findAllTopicsInBookmark(MemberFixture.createUser(otherMember));
-
-        assertThat(topicsInBookmark).hasSize(2);
-        assertThat(topicsInBookmark).extractingResultOf("topicId")
-                .containsExactlyInAnyOrder(topic1.getId(), topic2.getId());
-
     }
 
     private List<Long> getTopicsWithPermission(Member member) {
