+++ backend/src/test/java/com/mapbefine/mapbefine/member/presentation/MemberIntegrationTest.java
@@ -11,8 +11,6 @@ import com.mapbefine.mapbefine.location.domain.LocationRepository;
 import com.mapbefine.mapbefine.member.MemberFixture;
 import com.mapbefine.mapbefine.member.domain.Member;
 import com.mapbefine.mapbefine.member.domain.MemberRepository;
-import com.mapbefine.mapbefine.member.domain.MemberTopicBookmark;
-import com.mapbefine.mapbefine.member.domain.MemberTopicBookmarkRepository;
 import com.mapbefine.mapbefine.member.domain.MemberTopicPermission;
 import com.mapbefine.mapbefine.member.domain.MemberTopicPermissionRepository;
 import com.mapbefine.mapbefine.member.domain.Role;
@@ -20,7 +18,6 @@ import com.mapbefine.mapbefine.member.dto.request.MemberCreateRequest;
 import com.mapbefine.mapbefine.member.dto.request.MemberTopicPermissionCreateRequest;
 import com.mapbefine.mapbefine.member.dto.response.MemberDetailResponse;
 import com.mapbefine.mapbefine.member.dto.response.MemberResponse;
-import com.mapbefine.mapbefine.member.dto.response.MemberTopicBookmarkResponse;
 import com.mapbefine.mapbefine.member.dto.response.MemberTopicPermissionDetailResponse;
 import com.mapbefine.mapbefine.member.dto.response.MemberTopicPermissionResponse;
 import com.mapbefine.mapbefine.pin.PinFixture;
@@ -59,9 +56,6 @@ class MemberIntegrationTest extends IntegrationTest {
 
     @Autowired
     private MemberTopicPermissionRepository memberTopicPermissionRepository;
-
-    @Autowired
-    private MemberTopicBookmarkRepository memberTopicBookmarkRepository;
 
     @Test
     @DisplayName("Topic 을 만든자가 특정 유저에게 권한을 준다.")
@@ -410,144 +404,6 @@ class MemberIntegrationTest extends IntegrationTest {
                 .usingRecursiveComparison()
                 .ignoringFieldsOfTypes(LocalDateTime.class)
                 .isEqualTo(List.of(TopicResponse.from(topic1), TopicResponse.from(topic2)));
-    }
-
-    @Test
-    @DisplayName("유저가 토픽을 즐겨찾기 목록에 추가한다.")
-    void addTopicInBookmark_Success() {
-        //given
-        Member creator = MemberFixture.create("member", "member@naver.com", Role.USER);
-        memberRepository.save(creator);
-
-        Topic topic = TopicFixture.createByName("topic1", creator);
-        topicRepository.save(topic);
-
-        Member otherUser =
-                MemberFixture.create("otherUser", "otherUse@naver.com", Role.USER);
-        memberRepository.save(otherUser);
-
-        String authHeader = Base64.encodeBase64String(
-                ("Basic " + otherUser.getMemberInfo().getEmail()).getBytes()
-        );
-
-        //when
-        ExtractableResponse<Response> response = given().log().all()
-                .header(AUTHORIZATION, authHeader)
-                .param("topicId", topic.getId())
-                .when().post("/members/bookmarks")
-                .then().log().all()
-                .extract();
-
-        //then
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-        assertThat(response.header("Location")).startsWith("/members/bookmarks/")
-                .isNotNull();
-    }
-
-    @Test
-    @DisplayName("유저의 즐겨찾기 토픽 목록을 조회한다.")
-    void findTopicsInBookmarks_Success() {
-        //given
-        Member creator = MemberFixture.create("member", "member@naver.com", Role.USER);
-        memberRepository.save(creator);
-
-        Topic topic1 = TopicFixture.createByName("topic1", creator);
-        Topic topic2 = TopicFixture.createByName("topic1", creator);
-        topicRepository.save(topic1);
-        topicRepository.save(topic2);
-
-        Member otherUser =
-                MemberFixture.create("otherUser", "otherUse@naver.com", Role.USER);
-        memberRepository.save(otherUser);
-
-        MemberTopicBookmark bookmarkedTopic1 =
-                MemberTopicBookmark.createWithAssociatedTopicAndMember(topic1, otherUser);
-        MemberTopicBookmark bookmarkedTopic2 =
-                MemberTopicBookmark.createWithAssociatedTopicAndMember(topic2, otherUser);
-
-        memberTopicBookmarkRepository.save(bookmarkedTopic1);
-        memberTopicBookmarkRepository.save(bookmarkedTopic2);
-
-        String authHeader = Base64.encodeBase64String(
-                ("Basic " + otherUser.getMemberInfo().getEmail()).getBytes()
-        );
-
-        //when
-        List<MemberTopicBookmarkResponse> response = given().log().all()
-                .header(AUTHORIZATION, authHeader)
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .when().get("/members/bookmarks")
-                .then().log().all()
-                .statusCode(HttpStatus.OK.value())
-                .extract()
-                .as(new TypeRef<>() {
-                });
-
-        //then
-        assertThat(response).hasSize(2)
-                .usingRecursiveComparison()
-                .ignoringFieldsOfTypes(LocalDateTime.class)
-                .isEqualTo(List.of(
-                        MemberTopicBookmarkResponse.from(bookmarkedTopic1),
-                        MemberTopicBookmarkResponse.from(bookmarkedTopic2))
-                );
-    }
-
-    @Test
-    @DisplayName("유저의 즐겨찾기 토픽을 삭제한다.")
-    void deleteTopicInBookmark_Success() {
-        //given
-        Member creator = MemberFixture.create("member", "member@naver.com", Role.USER);
-        memberRepository.save(creator);
-
-        Topic topic = TopicFixture.createByName("topic1", creator);
-        topicRepository.save(topic);
-
-        MemberTopicBookmark bookmarkedTopic =
-                MemberTopicBookmark.createWithAssociatedTopicAndMember(topic, creator);
-        memberTopicBookmarkRepository.save(bookmarkedTopic);
-
-        String authHeader = Base64.encodeBase64String(
-                ("Basic " + creator.getMemberInfo().getEmail()).getBytes()
-        );
-
-        //when then
-        given().log().all()
-                .header(AUTHORIZATION, authHeader)
-                .when().delete("/members/bookmarks/" + bookmarkedTopic.getId())
-                .then().log().all()
-                .statusCode(HttpStatus.NO_CONTENT.value());
-    }
-
-    @Test
-    @DisplayName("유저의 즐겨찾기 토픽을 전체 삭제한다.")
-    void deleteAllTopicsInBookmark_Success() {
-        //given
-        Member creator = MemberFixture.create("member", "member@naver.com", Role.USER);
-        memberRepository.save(creator);
-
-        Topic topic1 = TopicFixture.createByName("topic1", creator);
-        Topic topic2 = TopicFixture.createByName("topic1", creator);
-        topicRepository.save(topic1);
-        topicRepository.save(topic2);
-
-        MemberTopicBookmark bookmarkedTopic1 =
-                MemberTopicBookmark.createWithAssociatedTopicAndMember(topic1, creator);
-        MemberTopicBookmark bookmarkedTopic2 =
-                MemberTopicBookmark.createWithAssociatedTopicAndMember(topic2, creator);
-        memberTopicBookmarkRepository.save(bookmarkedTopic1);
-        memberTopicBookmarkRepository.save(bookmarkedTopic2);
-
-        String authHeader = Base64.encodeBase64String(
-                ("Basic " + creator.getMemberInfo().getEmail()).getBytes()
-        );
-
-        //when then
-        given().log().all()
-                .header(AUTHORIZATION, authHeader)
-                .when().delete("/members/bookmarks/")
-                .then().log().all()
-                .statusCode(HttpStatus.NO_CONTENT.value());
     }
 
 }
